+++ crates/sotf-plugins/crates/sotf-plugin-ambisonics/src/lib/types.rs
@@ -1,3 +1,3 @@
-pub use super::config::AmbisonicsDecoderConfig;
+pub use super::params::Params as AmbisonicsDecoderConfig;
 
 pub type AmbisonicsDecoderParams = AmbisonicsDecoderConfig;